refactor(timeline): consolidate edit-preview store subscriptions into useEditPreviewShifts
Move the six per-item preview-store subscriptions (linked-edit, rolling,
ripple, track-push, slip, slide) plus the slide neighbor lookups into a
single hook. The subscriptions still use the same useShallow selectors,
so per-clip re-renders only fire when this item's participation in the
active gesture changes — same as before — but the timeline-item host
file no longer has 200 lines of inline preview-store boilerplate.

index.tsx shrinks by another ~190 lines.
